presales-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic[email]==2.5.3
cloud-sql-python-connector[pg8000]==1.7.0
python-dotenv==1.0.0
google-auth==2.27.0